+++ config/app.php
@@ -189,5 +189,5 @@ return [
     | The only modification by laravel config
     |
     */
-    'version' => '1.3.1'
+    'version' => '1.4.0'
 ];